Move `start!` method to the `DbPoller` module and reorganize files
